Refactor to remove class and struct specific adapters
