--- src/TextTabulator.Adapter.Reflection/IReflectionTabulatorAdapter.cs
@@ -1,6 +1,0 @@
-﻿namespace TextTabulator.Adapter.Reflection
-{
-    public interface IReflectionTabulatorAdapter : ITabulatorAdapter
-    {
-    }
-}
