--- src/TextTabulator.Adapter.Reflection/StructReflectionTabulatorAdapter.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace TextTabulator.Adapter.Reflection
-{
-    public interface IStructReflectionTabulatorAdapter : IReflectionTabulatorAdapter
-    {
-    }
-
-    public class StructReflectionTabulatorAdapter<TStruct> : IStructReflectionTabulatorAdapter where TStruct : struct
-    {
-        public IEnumerable<string>? GetHeaderStrings()
-        {
-            throw new NotImplementedException();
-        }
-
-        public IEnumerable<IEnumerable<string>> GetValueStrings()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
